+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Entries are listed in reverse chronological order.
 
 * MSRV is now 1.88.0
 * Migrate `group` to `0.14`, `jubjub` to `0.11`, `rand_core` to `0.10`
+* Breaking: the `frost` feature was removed; the FROST ciphersuites will be
+  moved to the [`frost`](https://github.com/ZcashFoundation/frost) repository.
+* Added an `internal` feature which exposes internal functions required to e.g.
+  implement the FROST ciphersuites. It is not covered by SemVer guarantees.
 
 ## 0.5.2